restruct file connector CI tests
